Added Offer
The process of finding the most suitable campaign and reflecting it on the invoice has been completed.

+++ database/migrations/2023_02_23_175351_create_orders_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('product_id');
             $table->string('product_title');
-            $table->unsignedBigInteger('category_id');
+            $table->unsignedInteger('category_id');
             $table->string('category_title');
             $table->string('author');
             $table->unsignedBigInteger('quantity');
